(float)3.0 to 3.f

+++ src/VortFunc.c
@@ -138,15 +138,15 @@ static float g_gaussian_3D(float rho){
 	/* = 1 to 8sf for rho ~>6. Taylor expansion otherwise */
 	assert(rho >= 0 && "Rho should not be -ve");
 	const float pi = 3.14159265359f;
-	if(rho > (float)6.){
-		return (float)1.;
+	if(rho > 6.f){
+		return 1.f;
 	} else {
 		/* Approximate erf using Abramowitz and Stegan 1.7.26 */
 		float a1 = 0.3480242f, a2 = -0.0958798f, a3 = 0.7478556f, p = 0.47047f;
 		float rho_sr2 = rho / sqrtf(2);
-		float t = (float) 1. / (1 + p * rho_sr2);
+		float t = 1.f / (1.f + p * rho_sr2);
 		float erf = 1.f-t * (a1 + t * (a2 + t * a3)) * expf(-rho_sr2 * rho_sr2);
-		float term2 = rho * sqrtf(2 / pi) * expf(-rho_sr2 * rho_sr2);
+		float term2 = rho * sqrtf(2.f / pi) * expf(-rho_sr2 * rho_sr2);
 		return erf - term2;
 	}
 }
@@ -154,7 +154,7 @@ static float g_gaussian_3D(float rho){
 static float zeta_gaussian_3D(float rho){
 	assert(rho >= 0 && "Rho should not be -ve");
 	const float pi = 3.14159265359f;
-	return sqrtf(2 / pi) * expf(-rho * rho / 2);
+	return sqrtf(2.f / pi) * expf(-rho * rho / 2.f);
 }
 
 static void combined_gaussian_3D(float rho, float* g, float* zeta) {
